// src/types/geocoder.d.ts
export declare namespace geocoders {
    namespace Geocoder {
// ...
        interface ReGeocode {
            /**
             * 地址组成元素
             */
            addressComponent: ReGeocodeAddressComponent;
            /**
             * 格式化地址
             */
            formattedAddress: string;
            /**
             * 道路信息列表
             */
            roads: Road[];
            /**
             * 道路路口列表
             */
            crosses: Cross[];
            /**
             * 兴趣点列表
             */
            pois: ReGeocodePoi[];
            aois?: ReGeocodeAoi[] | undefined;
        }
        interface ReGeocodeResult {
            /**
             * 状态说明
             */
            info: string;
            /**
             * 逆地理编码结果
             */
            regeocode: ReGeocode;
        }
        interface BatchReGeocodeResult {
            /**
             * 状态说明
             */
            info: string;
            /**
             * 批量逆地理编码结果
             */
            regeocodes: ReGeocode[];
        }
        type SearchStatus = 'complete' | 'no_data' | 'error';
    }
// ...
}