// src/types/amapJs.d.ts
export declare namespace AMapJs {
// ...
    interface MapsEventOptions {
        lnglat: LngLat;
        pixel: Pixel;
        type: string;
        target: any;
    }
// ...
    interface ReGeocode {
        addressComponent: AddressComponent;
        formattedAddress: string;
        roads: Road[];
        crosses: Cross[];
        pois: ReGeocodePoi[];
    }

    interface ReGeocodePoi {
        id: string;
        name: string;
        type: string;
        tel: string;
        distance: number;
        direction: string;
        address: string;
        location: LngLat;
        businessArea: string;
    }

    interface GeocodeResult {
        info: string;
        geocodes: LngLat[];
        resultNum: number;
    }

    interface ReGeocodeResult {
        info: string;
        regeocode: ReGeocode;
    }
// ...
  }